Improved performance by removing the 'confirmed' cache invalidation
It's covered by the block invalidation, so that should be all we need.
